+++ CHANGELOG.md
@@ -175,3 +175,4 @@ was removed or the new one was added to folder
 2.01
 - Larger texture atlas error margin
 - Minor TexMan fixes
+- Tag section can return all existing tags for iteration